bruk delt feilkoder
